refactor: Make components import explicitly
